+++ src/components/layer-controller/LayerControllerSubscriber.js
@@ -7,12 +7,23 @@ import {
   createGenerateClassName,
 } from '@material-ui/core/styles';
 import { createZarrLoader, createOMETiffLoader } from '@hubmap/vitessce-image-viewer';
-
 import TitleInfo from '../TitleInfo';
-import LayerController from './LayerController';
+import RasterLayerController from './RasterLayerController';
+import VectorLayerController from './VectorLayerController';
 import ImageAddButton from './ImageAddButton';
 import {
-  RASTER_ADD, LAYER_REMOVE, CLEAR_PLEASE_WAIT, METADATA_REMOVE, LAYER_ADD, METADATA_ADD,
+  RASTER_ADD,
+  LAYER_REMOVE,
+  CLEAR_PLEASE_WAIT,
+  METADATA_REMOVE,
+  LAYER_ADD,
+  METADATA_ADD,
+  CELLS_SET_OPACITY,
+  CELLS_ADD,
+  CELLS_TURN_ON,
+  MOLECULES_ADD,
+  MOLECULES_SET_OPACITY,
+  MOLECULES_TURN_ON,
 } from '../../events';
 import { controllerTheme } from './styles';
 import { DEFAULT_LAYER_PROPS } from './constants';
@@ -83,6 +94,8 @@ function publishLayer({ loader, imageData, layerId }) {
 
 function LayerControllerSubscriber({ onReady, removeGridComponent, theme }) {
   const [imageOptions, setImageOptions] = useState(null);
+  const [cells, setCellsEvent] = useState(null);
+  const [molecules, setMoleculesEvent] = useState(null);
   const [layersAndLoaders, setLayersAndLoaders] = useState([]);
   const memoizedOnReady = useCallback(onReady, []);
 
@@ -113,8 +126,18 @@ function LayerControllerSubscriber({ onReady, removeGridComponent, theme }) {
       PubSub.publish(CLEAR_PLEASE_WAIT, 'raster');
     }
     memoizedOnReady();
-    const token = PubSub.subscribe(RASTER_ADD, handleRasterAdd);
-    return () => PubSub.unsubscribe(token);
+    const rasterAddtoken = PubSub.subscribe(RASTER_ADD, handleRasterAdd);
+    const cellsAddToken = PubSub.subscribe(
+      CELLS_ADD, () => setCellsEvent(CELLS_SET_OPACITY),
+    );
+    const moleculesAddToken = PubSub.subscribe(
+      MOLECULES_ADD, () => setMoleculesEvent(MOLECULES_SET_OPACITY),
+    );
+    return () => {
+      PubSub.unsubscribe(rasterAddtoken);
+      PubSub.unsubscribe(cellsAddToken);
+      PubSub.unsubscribe(moleculesAddToken);
+    };
   }, [memoizedOnReady]);
 
   const handleImageAdd = async (imageData) => {
@@ -132,7 +155,7 @@ function LayerControllerSubscriber({ onReady, removeGridComponent, theme }) {
   };
   const layerControllers = layersAndLoaders.map(({ layerId, imageData, loader }) => (
     <Grid key={layerId} item style={{ marginTop: '10px' }}>
-      <LayerController
+      <RasterLayerController
         layerId={layerId}
         imageData={imageData}
         handleLayerRemove={() => handleLayerRemove(layerId, imageData.name)}
@@ -141,6 +164,7 @@ function LayerControllerSubscriber({ onReady, removeGridComponent, theme }) {
       />
     </Grid>
   ));
+
   return (
     <TitleInfo
       title="Layer Controller"
@@ -149,6 +173,22 @@ function LayerControllerSubscriber({ onReady, removeGridComponent, theme }) {
     >
       <StylesProvider generateClassName={generateClassName}>
         <ThemeProvider theme={controllerTheme[theme]}>
+          <Grid item>
+            {cells ? (
+              <VectorLayerController
+                label="Cell Segmentations"
+                handleOpacityChange={v => PubSub.publish(CELLS_SET_OPACITY, v)}
+                handleToggleChange={v => PubSub.publish(CELLS_TURN_ON, v)}
+              />
+            ) : null}
+            {molecules ? (
+              <VectorLayerController
+                label="Molecules"
+                handleOpacityChange={v => PubSub.publish(MOLECULES_SET_OPACITY, v)}
+                handleToggleChange={v => PubSub.publish(MOLECULES_TURN_ON, v)}
+              />
+            ) : null}
+          </Grid>
           {layerControllers}
           <Grid item>
             <ImageAddButton